# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: Homebase API
  version: 0.1.0
  description: >
    API for Homebase household operations. Most `/api/v1` routes require a valid
    `homebase_session` cookie created by the auth flow.
servers:
  - url: /
...
paths:
...
  /auth/oauth/callback:
    get:
      tags: [Auth]
      summary: Complete OAuth/OIDC login
      operationId: oauthCallback
      security: []
      responses:
        "302":
          description: Sets the session cookie and redirects to the web app.
        "400":
          $ref: "#/components/responses/Error"
  /auth/dev-login:
    get:
      tags: [Auth]
      summary: Development login
      operationId: devLogin
      security: []
      responses:
        "302":
          description: Sets a development session cookie when enabled.
        "404":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: homebase_session
...
  responses:
...
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
...
security:
  - sessionCookie: []